Move validation to new file
