+++ curation_service/app.py
@@ -1,7 +1,7 @@
 import os
 import json
 import re
-from typing import Optional, Tuple
+from typing import Optional
 
 import boto3
 import click
@@ -14,6 +14,7 @@ from flask import Flask, request, jsonify, url_for, abort, Response, \
     render_template, Blueprint
 from jinja2 import Environment, ChoiceLoader
 
+from curation_service.validation import validation_funcs
 from indra.assemblers.html import HtmlAssembler
 from indra.assemblers.html.assembler import loader as indra_loader, \
     _format_stmt_text, _format_evidence_text
@@ -107,43 +108,6 @@ def _put_file(file_path, content):
 
 # Needs to match 'KEY1:VALUE1;KEY2:VALUE2;...'. Trailing ';' is optional.
 # Let keys be case-insensitive alphabet strings and values be any alphanumeric strings.
-signor_pattern = re.compile(r'^([a-zA-Z]+:[a-zA-Z0-9]+;)*([a-zA-Z]+:[a-zA-Z0-9]+)?$')
-
-
-def _validate_signor_comments(text) -> Tuple[bool, str]:
-    valid_keys = {
-        'CELL', 'TAXID', 'DIRECT', 'EFFECT', 'SENTENCE', 'MECHANISM', 'RESIDUE'
-    }
-    valid_str = f"'{', '.join(valid_keys)}'"
-    # Check if the comment has a valid syntax
-    m = signor_pattern.match(text)
-
-    # Pattern is invalid
-    if not m:
-        return (
-            False,
-            "Invalid syntax. Should be 'KEY1:VALUE1;KEY2:VALUE2;...', where each key "
-            f"is one of {valid_str}."
-        )
-
-    # Now test if the keys are valid
-    invalid_keys = []
-    for key_value in text.split(';'):
-        if not key_value:
-            # Skip empty strings e.g. from trailing ';'
-            continue
-        key, value = key_value.split(':', maxsplit=1)
-        if key.upper() not in valid_keys:
-            invalid_keys.append(key)
-    if invalid_keys:
-        return False, (f"Invalid key(s): '{', '.join(invalid_keys)}'. Must be one of"
-                  " {valid_str}.")
-    return True, ""
-
-
-validation_funcs = {
-    "signor": _validate_signor_comments
-}
 
 
 @ui_blueprint.route('/list', methods=['GET'])
